:recycle: refactor api
Moved methods from api to their own files and divided methods with multiple responsibilities.
