+++ src/services/platform/platform.ts
@@ -4,6 +4,8 @@ import { ThanksProps } from "../../actions/thanks/thanks"
 import { Message, ViewTypes } from "../../models/platform/message"
 import { Logger } from "../logger/logger"
 import { UserInfo } from "./slack/methods/get-user-info"
+import { SlackBody } from "../../models/platform/slack/body";
+import { Community } from "../../models/database/community";
 
 export type PlatformName = "slack"
 
@@ -19,6 +21,10 @@ export abstract class Platform {
     const errorMessage = `The ${platformName} platform is not implemented`
     Logger.onError(errorMessage)
     throw Error(errorMessage)
+  }
+
+  getPlatformProps = async (data, getProps) => {
+    return await getProps(this, data)
   }
 
   getTempUserData = (userId: string, key: string): string[] | undefined => {
@@ -40,11 +46,12 @@ export abstract class Platform {
     }
   }
 
+  abstract getCommunity: (body: SlackBody) => Community
   abstract getView: (view: ViewTypes, options: any | undefined) => Promise<Message>
 
   abstract sendMessage: (channelId: string, message: Message) => Promise<void>
   abstract updateMessage: (messageId: any, message: Message) => Promise<void>
-  
+
   abstract getCommunityMembers: (communityId: string) => Promise<string[]>
   abstract getMembersByChannelId: (channelId: string) => Promise<string[]>
   abstract getUserInfo: (userId: string) => Promise<UserInfo | undefined>
